family-app/backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
django-cors-headers>=4.3.0
python-dotenv>=1.0.0
psycopg[binary]>=3.1.0
pytest>=7.4.0
pytest-django>=4.7.0

